+++ library/libsoft1/libflow/libnet/3.app/net.vnc.c
@@ -2,7 +2,7 @@
 #define u16 unsigned short
 #define u32 unsigned int
 #define u64 unsigned long long
-u64 serve_vnc(u64 fd, u64 type, u8* buf, u64 len)
+u64 serve_vnc(u64 fd, u64 type, u8* buf, int len)
 {
 	//"RFB 003.008\n"	//12byte
 	return 0;
